Added arrows and score on game HUD. Also worked on the arrow for scoring.
arrow-moving.svg, right-button.tscn, and right_button.gd are all part of the moving right arrow. This arrow will go from the bottom of the screen to the top as the song plays and if the player presses the respective key and arrow intersects the correct arrow, the player either scores points or depletes opponents healthbar. Yet to implement the movement of the arrow and scoring mechanics.